서린-재수정

+++ src/api/Mypage/scrapAPI.js
@@ -2,6 +2,13 @@ const BASE_URL = 'https://www.yeogida.net';
 
 /* 스크랩 폴더 목록 조회 */
 export const getFolderData = async () => {
+    const token = localStorage.getItem('token'); // localStorage에서 토큰을 가져옴
+
+    if (!token) {
+        console.error('토큰이 없습니다!');
+        return [];
+    }
+
     try {
         const response = await fetch(`${BASE_URL}/mypage/scrap`, {
             headers: {
@@ -22,6 +29,13 @@ export const getFolderData = async () => {
 
 /* 스크랩 폴더 생성 */
 export const fetchFolderAdd = async (folderName) => {
+    const token = localStorage.getItem('token'); // localStorage에서 토큰을 가져옴
+
+    if (!token) {
+        console.error('토큰이 없습니다!');
+        return [];
+    }
+
     try {
         const response = await fetch(`${BASE_URL}/mypage/scrap/add`, {
             method: 'POST',
@@ -46,6 +60,13 @@ export const fetchFolderAdd = async (folderName) => {
 
 /* 스크랩 폴더 삭제 */
 export const fetchFolderDelete = async (folderId) => {
+    const token = localStorage.getItem('token'); // localStorage에서 토큰을 가져옴
+
+    if (!token) {
+        console.error('토큰이 없습니다!');
+        return [];
+    }
+
     try {
         const response = await fetch(`${BASE_URL}/mypage/scrap/delete/${folderId}`, {
             method: 'DELETE',
@@ -66,6 +87,13 @@ export const fetchFolderDelete = async (folderId) => {
 
 /* 스크랩 폴더 이름 수정 */
 export const fetchFolderNameUpdate = async (folderId, folderName) => {
+    const token = localStorage.getItem('token'); // localStorage에서 토큰을 가져옴
+
+    if (!token) {
+        console.error('토큰이 없습니다!');
+        return [];
+    }
+
     try {
         const response = await fetch(`${BASE_URL}/mypage/scrap/rename/${folderId}`, {
             method: 'PUT',
@@ -90,6 +118,13 @@ export const fetchFolderNameUpdate = async (folderId, folderName) => {
 
 /* 특정 스크랩 폴더의 스크랩 목록 조회 */
 export const getScrapData = async (folderId) => {
+    const token = localStorage.getItem('token'); // localStorage에서 토큰을 가져옴
+
+    if (!token) {
+        console.error('토큰이 없습니다!');
+        return [];
+    }
+
     try {
         const response = await fetch(`${BASE_URL}/mypage/scrap/${folderId}`, {
             headers: {
@@ -110,6 +145,13 @@ export const getScrapData = async (folderId) => {
 
 /* 특정 스크랩 삭제 */
 export const fetchScrapDelete = async (folderId, scrapId) => {
+    const token = localStorage.getItem('token'); // localStorage에서 토큰을 가져옴
+
+    if (!token) {
+        console.error('토큰이 없습니다!');
+        return [];
+    }
+    
     try {
         const response = await fetch(`${BASE_URL}/mypage/scrap/${folderId}/delete/${scrapId}`, {
             method: 'DELETE',
